refactor: moved functions - Moved functions - Fixed imports - Removed duplicated file
