+++ src/builders/Button.ts
@@ -1,6 +1,5 @@
 import { type APIMessageComponentEmoji, type ButtonStyle, ComponentType, type APIButtonComponent } from '../types';
-import type { EmojiResolvable } from '../common';
-import { resolvePartialEmoji } from '../structures/extra/functions';
+import { resolvePartialEmoji, type EmojiResolvable } from '../common';
 
 /**
  * Represents a button component.
